optimizer/requirements.txt
fastapi
uvicorn
pymoo
numpy
python-dotenv
